pyproject.toml
[tool.poetry]
name = "foodify"
version = "0.1.0"
description = ""
authors = ["Luis Galdeano <lpgaldeano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^2.0.12"
psycopg2 = "^2.9.6"
requests = "^2.30.0"
pandas = "^2.0.1"
openfoodfacts = "^0.1.3"
functions = "^0.7.0"
webdriver-manager = "^3.8.6"
selenium = "^4.9.0"
bs4 = "^0.0.1"
opencv-python = "^4.7.0.72"
certifi = "^2023.5.7"
numpy = "^1.24.3"
opencv-contrib-python = "^4.7.0.72"
pyzbar = "^0.1.9"
pydub = "^0.25.1"
ffmpeg = "^1.4"
uvicorn = "^0.22.0"
fastapi = "^0.95.1"
sqlalchemy-utils = "^0.41.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

